+++ js/test/address/nodeunit/testaddress_CR.js
@@ -37,11 +37,11 @@ module.exports.testaddress_CR = {
 
     testParseAddressCRNormal: function(test) {
         test.expect(7);
-        var parsedAddress = new Address("Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia, San Rafael, San Rafael, 40501, COSTA RICA", {locale: 'es-CR'});
+        var parsedAddress = new Address("Señor Carlos Torres, Ca 15 Av 37 # 55\nHeredia, San Rafael\n40501\nCOSTA RICA", {locale: 'es-CR'});
         
         test.ok(typeof(parsedAddress) !== "undefined");
-        test.equal(parsedAddress.streetAddress, "Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia");
-        test.equal(parsedAddress.region, "San Rafael");
+        test.equal(parsedAddress.streetAddress, "Señor Carlos Torres, Ca 15 Av 37 # 55");
+        test.equal(parsedAddress.region, "Heredia");
         test.equal(parsedAddress.locality, "San Rafael");
         test.equal(parsedAddress.postalCode, "40501");
         test.equal(parsedAddress.country, "COSTA RICA");
@@ -51,25 +51,25 @@ module.exports.testaddress_CR = {
     
     testParseAddressCRNoZip: function(test) {
         test.expect(7);
-        var parsedAddress = new Address("Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia, San Rafael, San Rafael, 40501, COSTA RICA", {locale: 'es-CR'});
+        var parsedAddress = new Address("Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia, San Rafael, COSTA RICA", {locale: 'es-CR'});
         
         test.ok(typeof(parsedAddress) !== "undefined");
-        test.equal(parsedAddress.streetAddress, "Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia");
-        test.equal(parsedAddress.region, "San Rafael");
+        test.equal(parsedAddress.streetAddress, "Señor Carlos Torres, Ca 15 Av 37 # 55");
+        test.equal(parsedAddress.region, "Heredia");
         test.equal(parsedAddress.locality, "San Rafael");
         test.equal(parsedAddress.country, "COSTA RICA");
         test.equal(parsedAddress.countryCode, "CR");
-        test.equal(parsedAddress.postalCode, "40501");
+        test.ok(!parsedAddress.postalCode);
         test.done();
     },
     
     testParseAddressCRManyLines: function(test) {
         test.expect(7);
-        var parsedAddress = new Address("SEÑOR\nFEDERICO TERRAZAS ARIAS, Ca 15 Av 37 # 55\nHeredia, San Rafael, San Rafael\n40501, COSTA RICA", {locale: 'es-CR'});
+        var parsedAddress = new Address("SEÑOR\nFEDERICO TERRAZAS ARIAS, Ca 15 Av 37 # 55\nHeredia, San Rafael\n40501\nCOSTA RICA", {locale: 'es-CR'});
         
         test.ok(typeof(parsedAddress) !== "undefined");
-        test.equal(parsedAddress.streetAddress, "SEÑOR, FEDERICO TERRAZAS ARIAS, Ca 15 Av 37 # 55, Heredia");
-        test.equal(parsedAddress.region, "San Rafael");
+        test.equal(parsedAddress.streetAddress, "SEÑOR, FEDERICO TERRAZAS ARIAS, Ca 15 Av 37 # 55");
+        test.equal(parsedAddress.region, "Heredia");
         test.equal(parsedAddress.locality, "San Rafael");
         test.equal(parsedAddress.postalCode, "40501");
         test.equal(parsedAddress.country, "COSTA RICA");
@@ -79,11 +79,11 @@ module.exports.testaddress_CR = {
     
     testParseAddressCROneLine: function(test) {
         test.expect(7);
-        var parsedAddress = new Address("Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia, San Rafael, San Rafael, 40501, COSTA RICA", {locale: 'es-CR'});
+        var parsedAddress = new Address("Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia, San Rafael, 40501, COSTA RICA", {locale: 'es-CR'});
         
         test.ok(typeof(parsedAddress) !== "undefined");
-        test.equal(parsedAddress.streetAddress, "Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia");
-        test.equal(parsedAddress.region, "San Rafael");
+        test.equal(parsedAddress.streetAddress, "Señor Carlos Torres, Ca 15 Av 37 # 55");
+        test.equal(parsedAddress.region, "Heredia");
         test.equal(parsedAddress.locality, "San Rafael");
         test.equal(parsedAddress.postalCode, "40501");
         test.equal(parsedAddress.country, "COSTA RICA");
@@ -93,11 +93,11 @@ module.exports.testaddress_CR = {
     
     testParseAddressCRSuperfluousWhitespace: function(test) {
         test.expect(7);
-        var parsedAddress = new Address("Señor Carlos Torres, Ca 15 Av 37 # 55\n\t\n Heredia, San Rafael, San Rafael\t\n\n 40501, COSTA RICA  \n  \t\t\t", {locale: 'es-CR'});
+        var parsedAddress = new Address("Señor Carlos Torres, Ca 15 Av 37 # 55\n\t\n Heredia,    San Rafael\t\n\n 40501\n\nCOSTA RICA  \n  \t\t\t", {locale: 'es-CR'});
         
         test.ok(typeof(parsedAddress) !== "undefined");
-        test.equal(parsedAddress.streetAddress, "Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia");
-        test.equal(parsedAddress.region, "San Rafael");
+        test.equal(parsedAddress.streetAddress, "Señor Carlos Torres, Ca 15 Av 37 # 55");
+        test.equal(parsedAddress.region, "Heredia");
         test.equal(parsedAddress.locality, "San Rafael");
         test.equal(parsedAddress.postalCode, "40501");
         test.equal(parsedAddress.country, "COSTA RICA");
@@ -107,11 +107,11 @@ module.exports.testaddress_CR = {
     
     testParseAddressCRNoDelimiters: function(test) {
         test.expect(7);
-        var parsedAddress = new Address("SEÑOR Gabriel Garcia Marquez Ca 15 Av 37 # 55 Heredia San Rafael, San Rafael 40501, COSTA RICA", {locale: 'es-CR'});
+        var parsedAddress = new Address("SEÑOR Gabriel Garcia Marquez Ca 15 Av 37 # 55 Heredia San Rafael, 40501, COSTA RICA", {locale: 'es-CR'});
         
         test.ok(typeof(parsedAddress) !== "undefined");
-        test.equal(parsedAddress.streetAddress, "SEÑOR Gabriel Garcia Marquez Ca 15 Av 37 # 55 Heredia");
-        test.equal(parsedAddress.region, "San Rafael");
+        test.equal(parsedAddress.streetAddress, "SEÑOR Gabriel Garcia Marquez Ca 15 Av 37 # 55");
+        test.equal(parsedAddress.region, "Heredia");
         test.equal(parsedAddress.locality, "San Rafael");
         test.equal(parsedAddress.postalCode, "40501");
         test.equal(parsedAddress.country, "COSTA RICA");
@@ -121,11 +121,11 @@ module.exports.testaddress_CR = {
     
     testParseAddressCRSpecialChars: function(test) {
         test.expect(7);
-        var parsedAddress = new Address("SEÑOR, Gabriel García Márquez, SOCIEDAD DE ESCRITORES, Ca 15 Av 37 # 55, Heredia, San Rafael, San Rafael, 40501, COSTA RICA", {locale: 'es-CR'});
+        var parsedAddress = new Address("SEÑOR, Gabriel García Márquez, SOCIEDAD DE ESCRITORES, Ca 15 Av 37 # 55, Heredia, San Rafael, 40501, COSTA RICA", {locale: 'es-CR'});
         
         test.ok(typeof(parsedAddress) !== "undefined");
-        test.equal(parsedAddress.streetAddress, "SEÑOR, Gabriel García Márquez, SOCIEDAD DE ESCRITORES, Ca 15 Av 37 # 55, Heredia");
-        test.equal(parsedAddress.region, "San Rafael");
+        test.equal(parsedAddress.streetAddress, "SEÑOR, Gabriel García Márquez, SOCIEDAD DE ESCRITORES, Ca 15 Av 37 # 55");
+        test.equal(parsedAddress.region, "Heredia");
         test.equal(parsedAddress.locality, "San Rafael");
         test.equal(parsedAddress.postalCode, "40501");
         test.equal(parsedAddress.country, "COSTA RICA");
@@ -135,13 +135,13 @@ module.exports.testaddress_CR = {
     
     testParseAddressCRFromUS: function(test) {
         test.expect(7);
-        var parsedAddress = new Address("Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia, San Rafael, San Rafael, 40501, COSTA RICA", {locale: 'en-US'});
+        var parsedAddress = new Address("Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia, San Rafael, 40501, COSTA RICA", {locale: 'en-US'});
         
         // the country name is in English because this address is for a contact in a US database
         
         test.ok(typeof(parsedAddress) !== "undefined");
-        test.equal(parsedAddress.streetAddress, "Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia");
-        test.equal(parsedAddress.region, "San Rafael");
+        test.equal(parsedAddress.streetAddress, "Señor Carlos Torres, Ca 15 Av 37 # 55");
+        test.equal(parsedAddress.region, "Heredia");
         test.equal(parsedAddress.locality, "San Rafael");
         test.equal(parsedAddress.postalCode, "40501");
         test.equal(parsedAddress.country, "COSTA RICA");
@@ -152,14 +152,15 @@ module.exports.testaddress_CR = {
     testFormatAddressCR: function(test) {
         test.expect(1);
         var parsedAddress = new Address({
-            streetAddress: "Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia",
+            streetAddress: "Señor Carlos Torres, Ca 15 Av 37 # 55",
+            region: "Heredia",
             locality: "San Rafael",
-            region: "San Rafael",
             country: "COSTA RICA",
+            postalCode: "40501",
             countryCode: "CR"
         }, {locale: 'es-CR'});
         
-        var expected = "Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia\nSan Rafael, San Rafael\nCOSTA RICA";
+        var expected = "Señor Carlos Torres, Ca 15 Av 37 # 55\nHeredia, San Rafael\n40501\nCOSTA RICA";
         var formatter = new AddressFmt({locale: 'es-CR'});
         test.equal(formatter.format(parsedAddress), expected);
         test.done();
@@ -168,17 +169,17 @@ module.exports.testaddress_CR = {
     testFormatAddressCRFromUS: function(test) {
         test.expect(1);
         var parsedAddress = new Address({
-            streetAddress: "Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia",
+            streetAddress: "Señor Carlos Torres, Ca 15 Av 37 # 55",
             country: "COSTA RICA",
-            region: "San Rafael",
+            region: "Heredia",
             locality: "San Rafael",
+            postalCode: "40501",
             countryCode: "CR"
         }, {locale: 'en-US'});
         
-        var expected = "Señor Carlos Torres, Ca 15 Av 37 # 55, Heredia\nSan Rafael, San Rafael\nCOSTA RICA";
+        var expected = "Señor Carlos Torres, Ca 15 Av 37 # 55\nHeredia, San Rafael\n40501\nCOSTA RICA";
         var formatter = new AddressFmt({locale: 'en-US'});
         test.equal(formatter.format(parsedAddress), expected);
         test.done();
     }
-    
 };