+++ docs/ReleaseNotes.md
@@ -12,6 +12,14 @@ New Features:
     * Updated Name of the Day and Months part. If normal and standAlone are different, It displays both.
 * Modified ilib.getTimeZone() to find the time zone from the Intl object if it is available. If not, it falls back
 to the previous behaviour of checking environment variables.
+* Added AddressFmt.getFormatInfo() method for returning info that a UI would need to construct an address input form
+    * Return info about the correct form elements for an address in the given country, and return them in the right 
+    order
+    * Return localized UI labels for the various form elements
+    * Returns the localized list of regions for the address form if addresses in that country
+    typically include the region
+    * Returns a regular expression string to use for validation if the form element can be validated with a regexp
+    * Returns the localized list of countries
 
 Bug Fixes:
 * Fixed an ar-IQ currency symbol.